.requirements.txt
psycopg2
xmpppy
